requirements.txt
# Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Video generation
moviepy>=2.0.0
Pillow>=10.0.0

# Audio / voiceover
elevenlabs>=1.0.0

# Social media publishers
atproto>=0.0.46        # Bluesky AT Protocol
tweepy>=4.14.0          # Twitter/X API v2
requests>=2.31.0        # HTTP for Threads/Reels/TikTok

# Cloud storage (S3-compatible: Cloudflare R2)
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0            # CLI
rich>=13.0.0            # Pretty terminal output
